починил добавление картинок в превью фильтров и окна
