+++ js/form-validation.js
@@ -1,38 +1,15 @@
 import { isUniqueArr } from './data.js';
-import { startUpload } from './form-upload.js';
 
 const SubmitButtonText = {
 	IDLE: 'Опубликовать',
 	SUBMITTING: 'Отправляю...',
 };
-const FILE_TYPES = ['jpg', 'jpeg', 'png'];
 
 const form = document.querySelector('.img-upload__form');
 const inputHashtag = form.querySelector('input[name="hashtags"]');
 const submitButton = form.querySelector('.img-upload__submit');
-const fileField = form.querySelector('.img-upload__input');
-const photoPreview = form.querySelector('.img-upload__preview img');
-const effectsPrewiev = form.querySelector('.effects__preview ');
 
 let hashtagArray = [];
-
-const isValidType = (file) => {
-	const fileName = file.name.toLowerCase();
-	return FILE_TYPES.some((it) => fileName.endsWith(it));
-};
-
-//не могу понять куда экспортировать ее
-const onFileInputChange = () => {
-	const file = fileField.files[0];
-
-	if (file && isValidType(file)) {
-		photoPreview.src = URL.createObjectURL(file);
-		effectsPrewiev.forEach((preview) => {
-			preview.style.backgroundImage = `url('${photoPreview.src}')`;
-		});
-	}
-	startUpload();
-};
 
 const pristine = new Pristine(form, {
 	classTo: 'img-upload__field-wrapper',
